Add files via upload
Develop a script and record a video presenting your findings to the CEO and CMO based on the four questions they asked and the visuals you created in the previous tasks.

You can use your work or the model answer from the previous task to develop your presentation.

When writing your script, you should speak about your entire process, including the initial data load and clean-up steps so that your leaders know you’ve done your due diligence in providing error-free analysis. 
Data analysis provides heaps of information but remember to focus on the information that is most important to your leaders.

Your video should be approximately 5 minutes in length. You can record your video in the browser or upload a video from your device below.

Remember, this is not just a read out of your visuals, you’re presenting to the CEO and CMO and need to provide context for each scenario and how it relates to their questioning.